README.md: examples: explicitly close client
While playing around with the [Fauna tutorial](https://fauna.com/tutorials/crud#getting-started) I noticed the tutorial code, which I was basing off of the samples in the JVM client readme file, was hanging both in SBT and running inside IntelliJ.  After a bit of digging I saw it's because we are not explicitly closing the FaunaClient, so we end up with some leftover netty threads sitting in `kevent0()`.  

Just to make the example more explicit, this patch simply adds the call to close.  An alternative would be to wrap all the example code in a `sessionWith { ... }` function but this seemed like a less invasive change.